Move MemoryStore into pkg/keyring/testing.go
Committed-By-Agent: claude
